reetech-group-assest/objectDectect/requirements.txt
opencv-python==4.9.0.80
ultralytics==8.2.7
numpy==1.26.4
PyYAML==6.0.1